reorg and add optional parameter to set up LSST Sci Pipelines

+++ nersc/setup_sn_all.sh
@@ -1,24 +1,71 @@
-# Feb 2020: install SNANA on Cori
-# Apr 25 2020: load root
-# Aug 14 2020: load cfitsio & gsl
-# Feb 22 2021: R.Kessler - update & add ENVs for SNANA
-# Mar 05 2021: define few things for CosmoMC (installed by Vivian)
-#
-module unload python
-module unload PrgEnv-intel/6.0.5
-module load PrgEnv-gnu/6.0.5
-module swap gcc gcc/9.3.0
-module rm craype-network-aries
-module rm cray-libsci
-module unload craype
-module load cfitsio/3.47
-module add gsl
-module load root/6.18.00-py3
-module load intel/19.1.3.304  # for CosmoMC (Mar 5 2021)
-export CC=gcc
 
-# add SNANA to path
+# Jan 27 2022: HK update to optionally setup LSST Sci Pipelines
+# Mar 05 2021: define few things for CosmoMC (installed by Vivian)
+# Feb 22 2021: R.Kessler - update & add ENVs for SNANA
+# Aug 14 2020: load cfitsio & gsl
+# Apr 25 2020: load root
+# Feb 2020: install SNANA on Cori
+#
+
+shopt -s nocasematch
+
 export SN_GROUP='/global/cfs/cdirs/lsst/groups/SN'
+
+# setup without LSST Science Pipelines
+if [ -z "$1" ] 
+then
+  module unload python
+  module unload PrgEnv-intel/6.0.5
+  module load PrgEnv-gnu/6.0.5
+  module swap gcc gcc/9.3.0
+  module rm craype-network-aries
+  module rm cray-libsci
+  module unload craype
+  module load cfitsio/3.47
+  module load gsl
+  module load root/6.18.00-py3
+  module load intel/19.1.3.304  # for CosmoMC (Mar 5 2021)
+  export CC=gcc
+
+  export COSMOMC_DIR="$SN_GROUP/CosmoMCBBC"
+  export PATH=$PATH:${COSMOMC_DIR}
+
+  # Set up SN python
+  export LSST_INST_DIR=/global/common/software/lsst/common/miniconda
+  export SN_PYTHON_VER=sn-py
+  module unload python
+  unset PYTHONHOME
+  unset PYTHONPATH
+  export PYTHONNOUSERSITE=' '
+
+  # Just in case GCRCatalogs is installed
+  export DESC_GCR_SITE='nersc'
+
+  source $LSST_INST_DIR/$SN_PYTHON_VER/etc/profile.d/conda.sh
+  conda activate root
+  OUTPUTPY="$(which python)"
+  echo Now using "${OUTPUTPY}"
+
+# Setup with LSST Science Pipelines
+elif [[ "$1" == "lsst" ]];
+then
+  echo "Setting up SN env with LSST Science Pipelines"
+
+  module load cfitsio/3.47
+  module load root/6.18.00-py3
+
+  source /global/common/software/lsst/cori-haswell-gcc/stack/setup_any_stack.sh w_2021_40
+
+  export PYTHONPATH=$PYTHONPATH:/global/common/software/lsst/common/miniconda/sn-sprint-oct21/cvmfs/sw.lsst.eu/linux-x86_64/lsst_distrib/w_2021_40/conda/miniconda3-py38_4.9.2/envs/lsst-scipipe-0.7.0-ext/lib/python3.8/site-packages
+
+
+else
+  echo $1 "is an invalid option, please provide no parameters or use lsst to set up the LSST Science Pipelines"
+fi
+
+
+# Set up for all cases
+
 export SNANA_DIR="$SN_GROUP/snana/SNANA"
 export SNDATA_ROOT="$SN_GROUP/snana/SNDATA_ROOT"
 export SNANA_TESTS="$SN_GROUP/snana/SNANA_TESTS"
@@ -37,30 +84,9 @@ export PIPPIN_DIR="$SN_GROUP/Pippin"
 export SBATCH_TEMPLATES="$SNANA_LSST_ROOT/SBATCH_TEMPLATES"
 export SNANA_DEBUG="$SNANA_LSST_USERS/kessler/debug"
 
-export COSMOMC_DIR="$SN_GROUP/CosmoMCBBC"
-export PATH=$PATH:${COSMOMC_DIR}
-
-export PATH=$PATH:${SNANA_DIR}/bin:${SNANA_DIR}/util:${PIPPIN_DIR}
-
 # Aug 24 2020 RK - silly hack for CFITSIO
 export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:$CFITSIO_DIR/lib
 
-# Set up SN python
-export LSST_INST_DIR=/global/common/software/lsst/common/miniconda
-export SN_PYTHON_VER=sn-py
-module unload python
-unset PYTHONHOME
-unset PYTHONPATH
-export PYTHONNOUSERSITE=' '
-
-# Just in case GCRCatalogs is installed
-export DESC_GCR_SITE='nersc'
-
-source $LSST_INST_DIR/$SN_PYTHON_VER/etc/profile.d/conda.sh
-conda activate root
-OUTPUTPY="$(which python)"
-echo Now using "${OUTPUTPY}"
+export PATH=$PATH:${SNANA_DIR}/bin:${SNANA_DIR}/util:${PIPPIN_DIR}
 
 export HDF5_USE_FILE_LOCKING=FALSE
-
-
